-- if pointer has raw qualify, construct it as raw pointer type -- AS gets from function rather than hard code -- use raw as second level key for pointer cache -- judge raw by pointer type's attribute rather than qualify -- it is used to fix function type bug, function type will skip qualify

+++ clang/lib/CodeGen/CodeGenTypes.h
@@ -85,7 +85,7 @@ class CodeGenTypes {
   /// This map keeps cache of llvm::Types and maps clang::Type to
   /// corresponding llvm::Type.
   llvm::DenseMap<const Type *, llvm::Type *> TypeCache;
-  llvm::DenseMap<const Type *, llvm::DenseMap<uint64_t, llvm::Type *>> PointerTypeCache;
+  llvm::DenseMap<const Type *, llvm::DenseMap<unsigned int, llvm::Type *>> PointerTypeCache;
 
   llvm::DenseMap<const Type *, llvm::Type *> RecordsWithOpaqueMemberPointers;
